keep qwen3 omni veomni dpo path
Remove the non-veomni offline trainer path so this branch only carries the Omni-Preference data flow and VeOmni VLM DPO smoke coverage.
