--- tests/special_e2e/qwen3_omni_thinker_only_mllm_smoke.yaml
@@ -1,52 +1,0 @@
-# 2-GPU Thinker-only stage config for the Qwen3-Omni multimodal DPO smoke test.
-#
-# This mirrors qwen3_omni_thinker_only_smoke.yaml, but allows one image and one
-# video per prompt so the offline MLLM DPO path exercises visual inputs.
-stage_args:
-  - stage_id: 0
-    runtime:
-      devices: "0,1"
-    engine_args:
-      # ── Stage identity ──
-      model_stage: thinker
-      model_arch: Qwen3OmniMoeThinkerForConditionalGeneration
-      worker_type: ar
-      scheduler_cls: vllm_omni.core.sched.omni_ar_scheduler.OmniARScheduler
-      engine_output_type: text
-      hf_config_name: thinker_config
-
-      # ── Parallelism & backend ──
-      tensor_parallel_size: 2
-      distributed_executor_backend: "mp"
-
-      # ── Multimodal smoke inputs ──
-      limit_mm_per_prompt:
-        image: 1
-        video: 1
-        audio: 0
-
-      # ── Memory & batching (top-level CLI args are stripped) ──
-      gpu_memory_utilization: 0.4
-      max_num_seqs: 8
-      max_num_batched_tokens: 8192
-      enable_chunked_prefill: true
-
-      # ── Model loading ──
-      dtype: bfloat16
-      load_format: safetensors
-      trust_remote_code: true
-      enable_prefix_caching: false
-      enforce_eager: true
-
-      # ── LoRA (required for adapter hot-loading) ──
-      enable_lora: true
-      max_lora_rank: 64
-      max_loras: 1
-
-      # ── verl integration ──
-      enable_sleep_mode: true
-      logprobs_mode: processed_logprobs
-      disable_log_stats: true
-    final_output: true
-    final_output_type: text
-    is_comprehension: true
