+++ tests/special_e2e/run_qwen3_omni_veomni_vlm_dpo_smoke.sh
@@ -139,4 +139,3 @@ else
 fi
 
 echo "Qwen3-Omni VeOmni VLM DPO smoke test passed."
-
