+++ verl_omni/pipelines/model_base.py
@@ -14,7 +14,7 @@
 
 import logging
 from abc import ABC, abstractmethod
-from typing import Any, Optional
+from typing import Optional
 
 import torch
 from diffusers import ModelMixin, SchedulerMixin
@@ -267,140 +267,3 @@ class VllmOmniPipelineBase:
         if pipeline_cls is None:
             return None
         return f"{pipeline_cls.__module__}.{pipeline_cls.__qualname__}"
-
-
-class OmniModelBase(ABC):
-    """Plug-in base class for omni model training adapters."""
-
-    _registry: dict[tuple[str, str], type["OmniModelBase"]] = {}
-
-    @classmethod
-    def register(cls, architecture: str, stage: str = "thinker"):
-        """Class decorator that registers a subclass for ``(architecture, stage)``."""
-
-        def decorator(subclass: type["OmniModelBase"]) -> type["OmniModelBase"]:
-            cls._registry[(architecture, stage)] = subclass
-            return subclass
-
-        return decorator
-
-    @classmethod
-    def get_class(cls, model_config) -> type["OmniModelBase"]:
-        """Return the registered subclass for ``(architecture, stage)``."""
-        architecture = model_config.get("architecture", None)
-        stage = model_config.get("model_stage", "thinker")
-        key = (architecture, stage)
-
-        if key not in cls._registry and model_config.get("external_lib", None) is not None:
-            from verl.utils.import_utils import import_external_libs
-
-            import_external_libs(model_config.external_lib)
-
-        try:
-            return cls._registry[key]
-        except KeyError:
-            registered = sorted(cls._registry.keys())
-            raise NotImplementedError(
-                f"No omni model registered for (architecture={architecture!r}, stage={stage!r}). "
-                f"Registered: {registered}. Set ``external_lib`` to load your implementation."
-            ) from None
-
-    @classmethod
-    @abstractmethod
-    def get_model_architecture(cls) -> type:
-        """Return the Hugging Face model class to register for training."""
-        pass
-
-    @classmethod
-    @abstractmethod
-    def get_model_loading_kwargs(cls, model_config) -> dict[str, Any]:
-        """Return architecture-specific model loading kwargs."""
-        pass
-
-    @classmethod
-    @abstractmethod
-    def get_strip_modules(cls, model_config) -> list[str]:
-        """Return submodule prefixes to strip before wrapping the trainable stage."""
-        pass
-
-    @classmethod
-    @abstractmethod
-    def configure_processor(cls, model_path: str, model_config):
-        """Load and configure the multimodal processor for this omni model."""
-        pass
-
-    @classmethod
-    @abstractmethod
-    def configure_tokenizer(cls, model_path: str, model_config):
-        """Load and configure the tokenizer for this omni model."""
-        pass
-
-    @classmethod
-    def get_frozen_parameter_patterns(cls, model_config) -> list[str]:
-        """Return regex patterns for parameters to freeze during training."""
-        return []
-
-    @classmethod
-    def configure_model(cls, module: torch.nn.Module, model_config) -> torch.nn.Module:
-        """Apply generic stage stripping and parameter freezing."""
-        import re
-
-        for submod_name in cls.get_strip_modules(model_config):
-            if hasattr(module, submod_name):
-                delattr(module, submod_name)
-
-        frozen_patterns = [re.compile(pattern) for pattern in cls.get_frozen_parameter_patterns(model_config)]
-        if frozen_patterns:
-            for name, parameter in module.named_parameters():
-                if any(pattern.search(name) for pattern in frozen_patterns):
-                    parameter.requires_grad = False
-        return module
-
-    @classmethod
-    def apply_model_patches(cls) -> None:
-        """Install import-time compatibility patches, if the adapter needs any."""
-        return
-
-    @classmethod
-    def prepare_training_inputs(cls, module: torch.nn.Module, model_config, micro_batch: TensorDict) -> dict[str, Any]:
-        """Build training forward kwargs from a micro-batch."""
-        return dict(micro_batch)
-
-
-class OmniRolloutPipelineBase(ABC):
-    """Registry base for vLLM-Omni rollout pipeline topologies."""
-
-    _registry: dict[str, type["OmniRolloutPipelineBase"]] = {}
-
-    @classmethod
-    def register(cls, model_type: str):
-        """Class decorator that registers a subclass for ``model_type``."""
-
-        def decorator(subclass: type["OmniRolloutPipelineBase"]) -> type["OmniRolloutPipelineBase"]:
-            cls._registry[model_type] = subclass
-            return subclass
-
-        return decorator
-
-    @classmethod
-    def get_class(cls, model_type: str) -> type["OmniRolloutPipelineBase"] | None:
-        """Return the registered rollout adapter for ``model_type``."""
-        return cls._registry.get(model_type)
-
-    @classmethod
-    @abstractmethod
-    def get_stage_config(cls, pipeline_mode: str = "thinker_only", num_gpus: int = 8) -> list[dict[str, Any]]:
-        """Generate a vLLM-Omni stage config list."""
-        pass
-
-    @classmethod
-    @abstractmethod
-    def get_deploy_config(cls, pipeline_mode: str = "thinker_only", num_gpus: int = 8) -> dict[str, Any]:
-        """Generate a full vLLM-Omni deploy config."""
-        pass
-
-    @classmethod
-    @abstractmethod
-    def get_pipeline_model_type(cls) -> str:
-        """Return the vLLM-Omni pipeline model type."""
-        pass
