--- verl_omni/pipelines/qwen3_omni/__init__.py
@@ -1,19 +1,0 @@
-# Copyright 2026 Bytedance Ltd. and/or its affiliates
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from . import omni_rollout_adapter, thinker_training_adapter
-from .omni_rollout_adapter import *  # noqa: F401,F403
-from .thinker_training_adapter import *  # noqa: F401,F403
-
-__all__ = omni_rollout_adapter.__all__ + thinker_training_adapter.__all__
