+++ verl_omni/workers/config/__init__.py
@@ -12,8 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from . import diffusion, omni
+from . import diffusion
 from .diffusion import *  # noqa: F401,F403
-from .omni import *  # noqa: F401,F403
 
-__all__ = list(diffusion.__all__) + list(omni.__all__)
+__all__ = list(diffusion.__all__)
